Fix bootstrap.peon build on Unices.
It adds the ability to consider the --output-repository-dir option with
the build of a whole src repository.

Maybe we should consider to have only one repository for all stuff (src
like bin). Then this patch will not be useful anymore and it will fix
the problem with the Build-Depend field for the src packages; see #87.