Rework xLog and busLog.
busLog can not be used anymore for simple logging. It's the
responsability of xLog. Then, if an xLog object is passed to
busLog, the xLog messages are sent via busLog too.

For example:
  var xLog = require ('xcraft-core-log') (moduleName);
  require ('xcraft-core-buslog') (xLog);

  xLog.err ('The error');

The 'The error' message is sent in the bus via busLog if necessary.
The implementation is based on the EventEmitter of node.